Use TextMate.executor for improved output
Previously these used “selection or document” for what to compile/run, which I changed to entire document, as there is a dedicated command to run the selection. The commands now save the document though still work for untitled documents.

The runner will hyperlink errors, but since coffee run the generated JavaScript, the line and column references are wrong.